2nd Commit
First attempt at doing a for loop to print the ticker and the yearly change in the summary table. Probably not going to work. Going to test it out and then look into what I need to change. Need to figure out the correct formula to get the Yearly Change to be the closing amount in the last row for a ticker symbol minus the opening amount.